+ Se reemplaza en motor todos los double por float + Se Implementa metodo de parseo de numero float en Numero + Se muestran los errores de bresenham

+++ tp1/Motor.h
@@ -104,7 +104,7 @@ class Motor
          * @param xReal El valor x real de la circunferencia
          * @param distancia El error absoluto con respecto a la circunferencia
          */
-        double calcularIncertidumbrePje(double xReal, double distancia);
+        float calcularIncertidumbrePje(float xReal, float distancia);
 };
 
 #endif // MOTOR_H
